Add P25 decoding
A patched version of the OP25 project is needed to install some needed
functions.  Testing was conducted with a RTLSDR device with an E4000
tuner.  Even with your ppm entered, at higher frequencies your device
may still be off by a small amount.  Set the filter to "Wide" if
"Normal" doesn't work.