Translate pandapter painter with 0.5.
This provides a workaround for the "fixed" line drawing routines introduced in
Qt 4.8 / 5.0

From http://stackoverflow.com/questions/16990326:

It is correct that the rasterisation rules were improved and made consistent
in Qt 5. They are now consistent with how we render antialiased graphics (which
was always the correct way). Aliased graphics was always off-by-0.5 in a broken
way in Qt 4 which caused all sort of issues. The docs should be updated
accordingly. All the bottom-right-shifting logic applies to the Qt4 compat mode
only.

The default cap style (how we draw end points) for QPen is Qt::SquareCap. This
means that the line becomes half a pixel longer in each direction, totalling to
one pixel which lands at the bottom. Set the cap style to FlatCap and the extra
pixel goes away as it should.