Prevent crash due to device errors during start.
If a device error occurs during startup, e.g. no supported device is attached
to the computer, gr-osmosdr constructor will insert a null_source instead. The
idea was that the application should be able to run using this null_source;
however, gqrx still crashed because the driver returned 0 as current sample
rate.

This patch adds checking for such cases and presents a warning dialog to the user.
Gqrx should still run without crashing if the user chooses to ignore the
warning.