Restore correct filter preset in combo box.
Previously, any filter setting coming as low/high cut pairs were shown as "User
(X k)" in the filter preset selector. With this change, we always check whether
a new filter setting corresponds to one of the presets, if yes, we select the
corresponding option in the combo box.

Tested with bookmarks and manual settings, but not with remote.