Filter offset/width: allow utilizing whole bandwidth
What filter offsets and widths are usable can hardly be expressed as a fixed
fraction of the available bandwidth (currenty 90%), so allow setting an offset
or a width that brings the used frequencies right up to the edges of the
spectrum. Operating under these conditions should be avoided whenever possible,
but leave this decision to the user.

While being here fix the maximum selectable filter width for BFM for devices
with a sample rate between 200 and 500 KSa/s.