Range constraints: take filter_offset and lnb_lo into account
When checking if a frequency is inside the allowed range, or when setting it
to the middle of this range, the range is given wrt. the hardware rf
frequency. The LNB LO and the filter offset have to be blanked out for these
calculations.

This change allows demodulating frequencies below or above the hardware rf
range, as long as the signals are inside the available bandwidth. It also fixes
an issue where invalid hardware frequencies could be applied when loading the
configuration file.

Also see: https://groups.google.com/d/topic/gqrx/fIZrzhOlFfc/discussion