Fetch hardware frequency range only when necessary
We have to call updateFrequencyRange() on every change of the filter offset or
the LNB LO, which can happen frequently. Until now this always leads to
fetching the hardware frequency range via an API call.

The available hardware frequency range should not change after configuring the
device so store the range locally and reuse it in updateFrequencyRange().