Fix the square button issue on OS X.
As it turns out, the reason for the square buttons on OS X is a feature.
The button corners change automatically when the button width becomes
smaller than 50 or the height becomes smaller than 30 pixels. So, in otder
to prevent the buttons become square shaped I set the minimum size to
50x32 pixels.