Bookmark recall
I think that this change also belongs to this pull request in a broader sense, since a small problem with the bookmarks has already been fixed.

So far there was a problem if an FFT offset was larger than the bookmark frequency, then the bookmark frequency could not be read back. For example, if the offset was 800 KHz and the bookmark frequency was 147 KHz, then it stayed at 800 KHz receiving frequency and not the correct 147 KHz. To do this, the value rx-> set_filter_offset and uiDockRxOpt-> setFilterOffset are set to zero and the bookmark frequency is equal to the hardware frequency. In addition, the FFT plotter is also set to 0 so that the bookmark frequency is displayed centered.