Mouse Wheel PageStep
I use a rtlsdr stick with the setting extended_mode = 1. With it I can set the gain LNA, Mix and IF manually. The range of values extends from 0 to 15. Usually you use the mouse wheel for this, but unfortunately the default in dockinputctl.cpp does not match. The range of values is queried and setPageStep is set, but unfortunately only incompletely.

The change: from a value range of 15 everything remains as before, below 15 setPageStep is actively set to step = 1 and does not stay with the usual computer default of 3. Now you can set in 1 db steps with the mouse wheel instead of 3 db before...

The 15 is probably a default in gr-osmocom, but of course it can be applied to any other value.