second try for #908
I reconsidered the whole thing and the changes I have made so far have been plain garbage.

The procedure was wrong from the start because it makes sense to switch to normal bandwidth at least when changing the demodulator. A user bandwidth for USB makes no sense with AM or FM. As far as I understand the code, filter shape and mode use the same slot and you can no longer tell who has emitted something.

Here is my new attempt to fix the "error" - Filterwidth gets stuck in the user even though it was switched to normal. In addition, only changing the demodulator and not longer changing the filter shape is switching to normal.

The user filter remains active and selectable as long as the demodulator mode is not changed. If it is changed, the user filter is set to zero and switched off until a new user bandwidth is set with the current demodulator mode.