Prevent audio delay on lifftw plan generation
When ~/.gr_fftw_wisdom is missing, and gqrx is run for the first time,
there is noticeable audio delay, that grows while switching the FFT to
larger sizes. The audio is skipping too.
This patch prevents this from happening by unlocking data mutex and
locking separate FFT mutex for the time of libfftw plan generation.