Remove dependency on QRegExp
QRegExp has been removed from Qt 6 and replaced with QRegularExpression.
Although we could switch, it's just as easy to remove Gqrx's dependency
on regular expression processing.

In the RemoteControl class, we ean easily replace the regular expressions
with ordinary string processing functions. The MainWindow class tries to
extract a single-quoted portion of the device string for display in the
window title, but this doesn't seem useful (and could do more harm than
good). I've changed it to always display the raw device string.