Warn on audio device failure rather than abort.
In some cases, being able to hear the signal is not important, and a
system has no usable audio device. This patch removes the error dialog
and exit when no device is found. Instead, "NO AUDIO" is added to the
title. Note that the PortAudio case may still need some work. Only
PulseAudio has been tested.

Signed-off-by: Jeff Long <willcode4@gmail.com>