Add AB markers, measurement, and selection from current/peak fft
Markers A and B can be set from the filter frequency location (red line)
using the "->" buttons and cleared with the "x" buttons. When both A
and B are active, delta and center frequency are displayed.

Shift-click directly underneath the plot selects the span around the
cursor, at the dB level of the cursor. Ctrl-shift-click does the same
thing using the peak hold data, if enabled.

Markers can be dragged.

Also added a delta to filter frequency in the cursor text.

Signed-off-by: Jeff Long <willcode4@gmail.com>