Increase max fft size to 4M and reduce scope of locking in rx_fft
Larger ffts would cause overruns (testing done with a B205mini) and the
choke point was the write into the fft buffer. While the buffer data
structures need to be protected by a mutex, the fft execute() does not,
since it uses a separate buffer.

Increasing the number of threads had some effect on the achievable
fft rate for larger fft size, but was not the reason for the drops,
so settable nthreads can be left to another commit.

Signed-off-by: Jeff Long <willcode4@gmail.com>