When peak hold is selected, peak detect uses peak hold data
A sliding window is used for peak detection when using peak hold.
The original algorithm is still used when peak hold is off.

Peak detection is moved to after peak hold drawing, so the peak
hold data is available for use.

There was also an off-by-one in the drawing and detection of peaks.
While I don't understand the exact nature, adding 1 to the detected
bin seems to have fixed it.

Signed-off-by: Jeff Long <willcode4@gmail.com>