Scale FFT to dB/MHz
FFT power was scaled using 1/(fftsize^2). This works out approximately
for fftsize 1M, but was causing the power levels to shift at other
sizes.

The plotter shows the maximum value of all those mapped to a pixel on
the screen. This is useful, but does not represent the average power
for the bins mapped to that pixel. The code was probably written to
make the answer on the screen look correct in this situation, and show
the noise floor averaging down at higher fftsize.

Separate modes to show the average power in a fft bin, and the minimum
power, are the subjects of a separate commit.

Signed-off-by: Jeff Long <willcode4@gmail.com>